requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
importlib-metadata==4.8.1
install==1.3.4
Markdown==3.3.4
platformdirs==2.3.0
psycopg2==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
virtualenv==20.8.0
whitenoise==5.3.0
zipp==3.5.0
requests==2.28.0
schedule==1.1.0